make the landing page work
Together with @pietercolpaert we decided that there's no need for a `terms`, since the `@id` inside the `json` already, which is able to be linked though (after for example converting it to ttl). To simply use the dataset you use the URI as an identifier and thus don't really care about what the definition of it is.

This commit also moves into a `index` that shows some info, but it still needs to be fleshed out.